Removed the old build toolchain.
This was based on Eclipse, and I can't help to figure out what version
of Magick it was using. I managed to build it fine when I didn't have
the libmagick++-dev package installed, but once I installed it and got
CMake to work, this one wouldn't install or run.

The suggestion I read on the internet was using a fallback package with
the command `apt install graphicsmagick-libmagick-dev-compat
libgraphicsmagick++1-dev libmagick++-6-headers`.

This did the trick for the old version but I have no known mechanism to
rebuild the makefiles if I were to make changes to the code beyond just
minor adjustments