Touched up includes
Initially the code was giving me headaches because it wouldn't compile
when the assert.h was included from within the header files. It seems it
was not the reason, but still, removing a few unneeded includes.

The README had a few weird unicode characters that my editor didn't
appreciate.

Tested to compile properly on Debian.